refactor(form-generator): split page into components and fix indentation
- Split the monolithic form generator page into separate components (GeneratedForm, PromptTextarea, PublishDialog)
- Fix indentation issues in form.model.ts
- Remove unused import in useDashboard.ts
- Replace Loader with TablesLoader in formsTable.tsx
- Add type safety to generatedForm state in useFormGenerator.ts
